Redo web crawler in python & store results in mongo
Crawler redone in python for easier flow into the text analyzer, then
into the database.

`__main__.py` allows the directory to be run as a python module. It
will spawn a process which crawls websites, looking for new articles,
grabbing content, performing sentiment analysis, and storing the
results in mongo.